Move util to applications app
