+++ dataworkspace/dataworkspace/apps/datasets/utils.py
@@ -1,5 +1,3 @@
-import boto3
-from django.conf import settings
 from django.http import Http404
 from django.shortcuts import get_object_or_404
 
@@ -94,18 +92,3 @@ while (!dbHasCompleted(res)) {{
     sql_snippet = f"""SELECT * FROM "{schema}"."{table_name}" LIMIT 50"""
 
     return {"python": python_snippet, "r": r_snippet, "sql": sql_snippet}
-
-
-def get_quicksight_dashboard_name_url(dashboard_id):
-    account_id = boto3.client('sts').get_caller_identity().get('Account')
-    dashboard_name = boto3.client('quicksight').describe_dashboard(
-        AwsAccountId=account_id, DashboardId=dashboard_id, AliasName='$PUBLISHED'
-    )['Dashboard']['Name']
-    dashboard_url = boto3.client('quicksight').get_dashboard_embed_url(
-        AwsAccountId=account_id,
-        DashboardId=dashboard_id,
-        IdentityType='QUICKSIGHT',
-        UserArn=settings.QUICKSIGHT_DATASOURCE_USER_ARN,
-    )['EmbedUrl']
-
-    return dashboard_name, dashboard_url
